# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express API with Swagger",
    "version": "1.0.0"
  },
  "paths": {
    "/auth/register": {
      "get": {
        "summary": "Display registration page",
        "responses": {
          "200": {
            "description": "Successfully displays the registration page",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Redirects to login page after successful registration",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "example": "/auth/login"
                }
              }
            }
          },
          "500": {
            "description": "Error occurred during registration",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "get": {
        "summary": "Display login page",
        "responses": {
          "200": {
            "description": "Successfully displays the login page",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "User login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Redirects to home page after successful login",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "example": "/"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username or password",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error occurred during login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/mobile-login": {
      "post": {
        "summary": "Mobile user login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/specialties": {
      "get": {
        "summary": "List all specialties",
        "responses": {
          "200": {
            "description": "List of all specialties",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Specialty"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error fetching specialties",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string// for Kotlin/Swift"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/doctor/availability": {
      "get": {
        "summary": "Display the doctor availability form",
        "operationId": "getDoctorAvailability",
        "security": [
          {
            "sessionAuth": []
          },
          {
            "doctorRole": []
          }
        ],
        "responses": {
          "200": {
            "description": "Doctor availability form page rendered successfully",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Access denied for non-doctors",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Access denied. Only doctors can perform this action."
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Post a doctor's availability",
        "operationId": "postDoctorAvailability",
        "security": [
          {
            "sessionAuth": []
          },
          {
            "doctorRole": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date",
                    "description": "Date for which availability is being posted."
                  },
                  "availableHours": {
                    "type": "string",
                    "description": "Hours during which the doctor is available."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "303": {
            "description": "Redirects after posting availability",
            "headers": {
              "Location": {
                "description": "URL to redirect to after successful post",
                "schema": {
                  "type": "string",
                  "example": "/"
                }
              }
            }
          },
          "500": {
            "description": "Error occurred while posting availability",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Error while posting availability. Please try again later."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "specialty": {
            "type": "string"
          },
          "localization": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "Specialty": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionToken",
        "description": "Session token used for authenticating API requests"
      },
      "doctorRole": {
        "type": "apiKey",
        "in": "header",
        "name": "Role",
        "description": "Ensure the user role is 'doctor'"
      }
    }
  },
...
}